ai-models/requirements.txt
# FastAPI and async
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Basic Pitch for music transcription
basic-pitch==0.3.2

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.3
scipy==1.11.4

# Database and Redis
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.25.2

# Video processing (for YouTube)
yt-dlp==2023.12.30

# Music theory and MIDI
music21==9.1.0
midiutil==1.2.1
pretty-midi==0.2.10

# ML/AI dependencies
torch>=2.0.0
torchaudio>=2.0.0
tensorflow>=2.14.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
pylint==3.0.3
mypy==1.8.0